Sign all session cookies
Fixes #10

+++ spec/describe-session-store.js
@@ -1,15 +1,16 @@
 require('./helper');
-var RSVP = require('rsvp');
 module.exports = describeSessionStore;
 
 function describeSessionStore(store, skip) {
   if (!skip) {
     beforeEach(function () {
-      return store.purge();
+      if (typeof store.purge === 'function')
+        return store.purge();
     });
 
     after(function () {
-      return store.destroy();
+      if (typeof store.destroy === 'function')
+        return store.destroy();
     });
   }
 
@@ -18,13 +19,13 @@ function describeSessionStore(store, skip) {
   desc('when there is no session with a given value', function () {
     var session;
     beforeEach(function () {
-      return RSVP.resolve(store.load('fake-value')).then(function (newSession) {
+      return store.load('fake-value').then(function (newSession) {
         session = newSession;
       });
     });
 
     it('returns an empty object', function () {
-      assert.deepEqual(session, {});
+      expect(session).toEqual({});
     });
   });
 
@@ -32,41 +33,41 @@ function describeSessionStore(store, skip) {
     var value;
     beforeEach(function () {
       var session = { count: 1 };
-      return RSVP.resolve(store.save(session)).then(function (newValue) {
+      return store.save(session).then(function (newValue) {
         value = newValue;
       });
     });
 
     it('can be retrieved using the opaque return value', function () {
-      return RSVP.resolve(store.load(value)).then(function (session) {
+      return store.load(value).then(function (session) {
         assert(session);
-        assert.equal(session.count, 1);
+        expect(session.count).toEqual(1);
       });
     });
   });
 
   desc('when it has a TTL', function () {
     beforeEach(function () {
-      store.ttl = 10;
+      store._ttl = 10;
     });
 
     afterEach(function () {
-      delete store.ttl;
+      delete store._ttl;
     });
 
     describe('and a session is not expired', function () {
       var value;
       beforeEach(function () {
         var session = { count: 1 };
-        return RSVP.resolve(store.save(session)).then(function (newValue) {
+        return store.save(session).then(function (newValue) {
           value = newValue;
         });
       });
 
       it('loads the session', function () {
-        return RSVP.resolve(store.load(value)).then(function (session) {
+        return store.load(value).then(function (session) {
           assert(session);
-          assert.equal(session.count, 1);
+          expect(session.count).toEqual(1);
         });
       });
     });
@@ -75,45 +76,45 @@ function describeSessionStore(store, skip) {
       var value;
       beforeEach(function () {
         var session = { count: 1 };
-        return RSVP.resolve(store.save(session)).then(function (newValue) {
+        return store.save(session).then(function (newValue) {
           value = newValue;
-          return delay(store.ttl);
+          return delay(store._ttl);
         });
       });
 
       it('loads a new session', function () {
-        return RSVP.resolve(store.load(value)).then(function (session) {
+        return store.load(value).then(function (session) {
           assert(session);
-          assert.deepEqual(session, {});
+          expect(session).toEqual({});
         });
       });
     });
 
-    describe('and a session is touched before it expires', function () {
+    describe('and a session is saved before it expires', function () {
       var value;
       beforeEach(function () {
         var session = { count: 1 };
-        return RSVP.resolve(store.save(session)).then(function () {
-          return delay(store.ttl / 2).then(function () {
-            return RSVP.resolve(store.touch(session)).then(function () {
-              return RSVP.resolve(store.save(session)).then(function (newValue) {
-                value = newValue;
-                return delay(store.ttl / 2);
-              });
+        return store.save(session).then(function () {
+          return delay(store._ttl / 2).then(function () {
+            return store.save(session).then(function (newValue) {
+              value = newValue;
+              return delay(store._ttl / 2);
             });
           });
         });
       });
 
       it('loads the session', function () {
-        return RSVP.resolve(store.load(value)).then(function (session) {
+        return store.load(value).then(function (session) {
           assert(session);
-          assert.equal(session.count, 1);
+          expect(session.count).toEqual(1);
         });
       });
     });
   });
 }
+
+var RSVP = require('rsvp');
 
 function delay(ms) {
   var deferred = RSVP.defer();
